+++ yazi-actor/src/app/report.rs
@@ -1,9 +1,7 @@
-use std::sync::Arc;
-
 use anyhow::Result;
 use yazi_adapter::ADAPTOR;
 use yazi_emulator::{EMULATOR, Mux};
-use yazi_macro::{act, succ};
+use yazi_macro::{act, log_if_err, succ};
 use yazi_proxy::AppProxy;
 use yazi_shared::data::Data;
 use yazi_term::event::Report as TermReport;
@@ -18,30 +16,38 @@ impl Actor for Report {
 	const NAME: &str = "report";
 
 	fn act(cx: &mut Ctx, report: Self::Form) -> Result<Data> {
-		let Some(term) = cx.term.as_mut() else { succ!() };
-		let old_light = term.probe.emulator.light();
-
-		term.probe.emulator.apply(&report);
-		EMULATOR.store(Arc::new(term.probe.emulator.clone()));
-
-		if report.is_color_scheme()
-			&& old_light.zip(term.probe.emulator.light()).is_some_and(|(a, b)| a != b)
-			&& !term.probe.needs_passthrough()
-		{
-			return act!(app:theme, cx);
-		} else if !report.is_da_1() {
+		if cx.term.is_none() {
 			succ!();
-		} else if !term.probe.needs_passthrough() {
-			ADAPTOR.resolve(&term.probe.emulator);
-			return act!(app:theme, cx);
 		}
 
-		let id = term.probe.id;
+		let old_light = EMULATOR.light();
+		EMULATOR.apply(&report);
+
+		if EMULATOR.light() != old_light {
+			log_if_err!(act!(app:theme, cx));
+		}
+
+		if !report.is_da_1() {
+			succ!();
+		} else if EMULATOR.needs_passthrough() {
+			succ!(Self::reprobe());
+		}
+
+		ADAPTOR.resolve(&EMULATOR);
+		if EMULATOR.light().is_none() {
+			log_if_err!(act!(app:theme, cx));
+		}
+
+		succ!();
+	}
+}
+
+impl Report {
+	fn reprobe() {
+		let id = EMULATOR.probe_id.get();
 		tokio::spawn(async move {
 			Mux::tmux_setup().await;
 			AppProxy::passthrough(id);
 		});
-
-		succ!();
 	}
 }