+++ yazi-config/src/lib.rs
@@ -2,12 +2,13 @@ yazi_macro::mod_pub!(keymap mgr open opener plugin popup preview tasks theme vfs
 
 yazi_macro::mod_flat!(icon inject layout mixing pattern platform preset priority selectable selector tests yazi);
 
-use std::io::{Read, Write};
+use std::io::Write;
 
 use anyhow::Context;
 use yazi_fs::Xdg;
 use yazi_macro::writef;
 use yazi_shim::{cell::{RoCell, SyncCell}, toml::{DeserializeOver, DeserializeOverWith}};
+use yazi_term::TERM;
 use yazi_tty::{TTY, sequence::SetSgr};
 
 use crate::theme::{Flavor, Theme};
@@ -18,13 +19,12 @@ pub static THEME: RoCell<Theme> = RoCell::new();
 pub static VFS: RoCell<vfs::Vfs> = RoCell::new();
 pub static LAYOUT: SyncCell<Layout> = SyncCell::new(Layout::default());
 
-pub fn init() -> anyhow::Result<()> {
+pub fn setup() -> anyhow::Result<()> {
 	if let Err(e) = try_init(true) {
 		wait_for_key(e)?;
 		try_init(false)?;
 	}
 
-	THEME.init(Preset::theme(false)?.reshape(false)?);
 	Ok(())
 }
 
@@ -32,20 +32,24 @@ fn try_init(merge: bool) -> anyhow::Result<()> {
 	let mut yazi = Preset::yazi()?;
 	let mut keymap = Preset::keymap()?;
 	let mut vfs = Preset::vfs()?;
+	let mut theme = Preset::theme(false)?;
 
 	if merge {
 		yazi = parse("yazi.toml", yazi.deserialize_over(&yazi::Yazi::read()?))?;
 		keymap = parse("keymap.toml", keymap.deserialize_over(&keymap::Keymap::read()?))?;
 		vfs = parse("vfs.toml", vfs.deserialize_over(&vfs::Vfs::read()?))?;
+		theme = parse("theme.toml", theme.deserialize_over(&Theme::read()?))?;
 	} else {
 		yazi = yazi.deserialize_over("")?;
 		keymap = keymap.deserialize_over("")?;
 		vfs = vfs.deserialize_over("")?;
+		theme = theme.deserialize_over("")?;
 	}
 
 	YAZI.init(yazi);
 	KEYMAP.init(keymap);
 	VFS.init(vfs);
+	THEME.init(theme.reshape(false)?);
 	Ok(())
 }
 
@@ -73,22 +77,23 @@ where
 }
 
 fn wait_for_key(e: anyhow::Error) -> anyhow::Result<()> {
-	let mut stdout = &mut *TTY.lockout();
+	let mut stdout = TTY.lockout();
 
-	writeln!(stdout, "{e}")?;
+	write!(stdout, "{e}\r\n")?;
 	if let Some(src) = e.source() {
-		writeln!(stdout, "\nCaused by:\n{src}")?;
+		write!(stdout, "\r\nCaused by:\r\n{src}\r\n")?;
 	}
 
 	writef!(
 		stdout,
-		"{}{}Press <Enter> to continue with preset settings...{}",
+		"{}{}Press any key to continue with preset settings...{}",
 		SetSgr::Reverse,
 		SetSgr::Bold,
 		SetSgr::Reset
 	)?;
 
-	TTY.reader().read_exact(&mut [0])?;
+	drop(stdout);
+	TERM.source.try_poll(None, |event| event.is_key())?;
 	Ok(())
 }